requirements.txt
Django==4.1.3
google_api_python_client==2.73.0
google_auth_oauthlib==0.8.0
pymongo==4.3.3
PyYAML==6.0
